// web/src/api/models/api-key.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiKey
 */
export interface ApiKey {
    /**
     * 
     * @type {string}
     * @memberof ApiKey
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKey
     */
    'key'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKey
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKey
     */
    'created_at'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKey
     */
    'updated_at'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiKey
     */
    'last_used_at'?: string;
}